Fix `attrs.field(converter=...)` annotations
Converters can currently be specified as `_ConverterType`, `list[_ConverterType]`, or `tuple[_ConverterType]`. The latter does not mean "tuple of `_ConverterType` instances" as was likely intended; it requires the tuple to have exactly one element. The correct form for a `tuple` with any number of instances is `tuple[_ConverterType, ...]`.

I was tempted to change the parameter's type by replacing both the `list` and `tuple` expressions with a unified `Sequence[_ConverterType]`, but that currently doesn't work at runtime.